requirements.txt
numpy==1.20.1
matplotlib==3.3.4
streamlit==0.85.1
pandas==1.2.4
Pillow==8.3.1
scikit_learn==0.24.2
